+++ sample/00-showcase/src/app.module.ts
@@ -1,0 +1,13 @@
+import { AiModule } from '@nest-native/ai-sdk';
+import { Module } from '@nestjs/common';
+import { ChatModule } from './chat/chat.module';
+
+@Module({
+  imports: [
+    AiModule.forRoot({
+      defaultHeaders: { 'x-powered-by-ai-sdk': 'nest-native' },
+    }),
+    ChatModule,
+  ],
+})
+export class AppModule {}